ai/docker/marker-api/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Celery and Redis
celery[redis]>=5.3.0
redis>=5.0.0
flower>=2.0.0

# Pydantic for data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Async support
aiofiles>=23.2.0
httpx>=0.26.0

# Image processing (for base64 encoding)
Pillow>=10.0.0

# Environment management
python-dotenv>=1.0.0

# Monitoring and logging
structlog>=24.1.0
